requirements.txt
insightface>=0.7.3
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
onnxruntime-gpu>=1.16.0
tensorflow-gpu>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
mediapipe>=0.10.0
matplotlib>=3.7.0
scikit-learn>=1.3.0